+++ SolutionItems/Documents/Release Notes.md
@@ -1,5 +1,13 @@
 ﻿### Release Notes
 
+**0.36**
+
+* Billion Dollar Question
+* ElementField - IndexType property was added
+There are two types; Aggressive (existing) & Passive
+Passive case covers CMRPs that only use Initial value (like Billion Dollar Question)
+* Content updates
+
 **0.35.2**
 
 * Content updates